Fixed warning on windows builds.

+++ base/vulkanMeshLoader.hpp
@@ -394,9 +394,9 @@ public:
 	{
 
 		std::vector<float> vertexBuffer;
-		for (int m = 0; m < m_Entries.size(); m++)
+		for (size_t m = 0; m < m_Entries.size(); m++)
 		{
-			for (int i = 0; i < m_Entries[m].Vertices.size(); i++)
+			for (size_t i = 0; i < m_Entries[m].Vertices.size(); i++)
 			{
 				// Push vertex data depending on layout
 				for (auto& layoutDetail : layout)
